fix: вынести логику статистики в отдельный сервис в event fix: вынести метод обновления полей в маппер в event
