+++ main-service/src/main/java/ru/practicum/mainservice/event/mapper/EventMapper.java
@@ -4,14 +4,19 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import ru.practicum.mainservice.category.mapper.CategoryMapper;
 import ru.practicum.mainservice.category.model.Category;
+import ru.practicum.mainservice.category.repository.CategoryRepository;
 import ru.practicum.mainservice.event.dto.EventFullDto;
 import ru.practicum.mainservice.event.dto.EventShortDto;
 import ru.practicum.mainservice.event.dto.NewEventDto;
+import ru.practicum.mainservice.event.dto.UpdateEventRequest;
 import ru.practicum.mainservice.event.model.Event;
 import ru.practicum.mainservice.event.model.State;
+import ru.practicum.mainservice.exception.exception.BadRequestException;
+import ru.practicum.mainservice.exception.exception.NotFoundException;
 import ru.practicum.mainservice.user.mapper.UserMapper;
 import ru.practicum.mainservice.user.model.User;
 
+import java.time.Duration;
 import java.time.LocalDateTime;
 
 @Component
@@ -75,4 +80,58 @@ public class EventMapper {
                 .views(event.getViews())
                 .build();
     }
+
+    public void updateEventFields(Event event, UpdateEventRequest updateEventRequest, StringBuilder updatedFieldsLog, CategoryRepository categoryRepository) {
+        if (updateEventRequest.getAnnotation() != null) {
+            event.setAnnotation(updateEventRequest.getAnnotation());
+            updatedFieldsLog.append("Annotation|");
+        }
+        if (updateEventRequest.getCategory() != null) {
+            event.setCategory(categoryRepository.findById(updateEventRequest.getCategory())
+                    .orElseThrow(() -> new NotFoundException(String.format("Category with id=%d not found", updateEventRequest.getCategory()))));
+            updatedFieldsLog.append("Category|");
+        }
+        if (updateEventRequest.getDescription() != null) {
+            event.setDescription(updateEventRequest.getDescription());
+            updatedFieldsLog.append("Description|");
+        }
+        if (updateEventRequest.getLocation() != null) {
+            event.setLocation(updateEventRequest.getLocation());
+            updatedFieldsLog.append("Location|");
+        }
+        if (updateEventRequest.getPaid() != null) {
+            event.setPaid(updateEventRequest.getPaid());
+            updatedFieldsLog.append("Paid|");
+        }
+        if (updateEventRequest.getParticipantLimit() != null) {
+            event.setParticipantLimit(updateEventRequest.getParticipantLimit());
+            updatedFieldsLog.append("ParticipantLimit|");
+        }
+        if (updateEventRequest.getRequestModeration() != null) {
+            event.setRequestModeration(updateEventRequest.getRequestModeration());
+            updatedFieldsLog.append("RequestModeration|");
+        }
+        if (updateEventRequest.getEventDate() != null) {
+            validateEventDate(updateEventRequest.getEventDate());
+            event.setEventDate(updateEventRequest.getEventDate());
+            updatedFieldsLog.append("EventDate|");
+        }
+        if (updateEventRequest.getTitle() != null) {
+            event.setTitle(updateEventRequest.getTitle());
+            updatedFieldsLog.append("Title|");
+        }
+    }
+
+    private void validateEventDate(LocalDateTime eventDate) {
+        if (!isStartDateValid(LocalDateTime.now(), eventDate, 2)) {
+            throw new BadRequestException("The date and time for which the event is scheduled cannot be earlier" +
+                    " than two hours from the current moment.");
+        }
+    }
+
+    private boolean isStartDateValid(LocalDateTime publicationDate, LocalDateTime startDate, int constraint) {
+        long hoursBetween = Duration.between(publicationDate, startDate).toHours();
+        return hoursBetween >= constraint;
+    }
+
 }
